fix for strange error with kodi texturedb

+++ lib/helpers/utils.py
@@ -456,11 +456,11 @@ def refresh_image(imagepath):
     connection = sqlite3.connect(dbpath, timeout=30, isolation_level=None)
     try:
         cache_image = connection.execute('SELECT cachedurl FROM texture WHERE url = ?', (imagepath,)).fetchone()
-        if cache_image:
+        if cache_image and isinstance(cache_image, (unicode, str)):
             if xbmcvfs.exists(cache_image):
                 xbmcvfs.delete("special://profile/Thumbnails/%s" % cache_image)
             connection.execute('DELETE FROM texture WHERE url = ?', (imagepath,))
-            connection.close()
+        connection.close()
     except Exception as exc:
         log_exception(__name__, exc)
     finally:
